+++ lib/modules/forms/forms_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: forms
-// Updated at: 2026-05-13 16:09:17
-// Commit: chore(forms): update dependencies and cleanup unused imports in forms
+// Updated at: 2026-05-14 12:05:48
+// Commit: feat(forms): add bi-directional event bus handler for forms
 
 class formsComponentState {
-  final String commitRef = '1109';
-  final String lastMessage = 'chore(forms): update dependencies and cleanup unused imports in forms';
-  final DateTime timestamp = DateTime.parse('2026-05-13T16:09:17+05:00');
+  final String commitRef = '1134';
+  final String lastMessage = 'feat(forms): add bi-directional event bus handler for forms';
+  final DateTime timestamp = DateTime.parse('2026-05-14T12:05:48+05:00');
 }